+++ index.js
@@ -1,4 +1,7 @@
 export { DynamixelController } from './src/DynamixelController.js';
 export { DynamixelDevice, Protocol2 } from './src/dynamixel/index.js';
 export { U2D2Connection, SerialConnection, WebSerialConnection } from './src/transport/index.js';
+export { AlarmManager } from './src/dynamixel/AlarmManager.js';
+export { MotorProfiles } from './src/dynamixel/MotorProfiles.js';
+export { Logger } from './src/utils/Logger.js';
 export * from './src/dynamixel/constants.js';